added useProductFilter hook
