+++ client/src/pages/products/components/ProductFilter.tsx
@@ -1,5 +1,5 @@
-import React, { useMemo, memo, useEffect } from 'react';
-import { useState } from 'react';
+import React from 'react';
+import { useProductFilter } from '../hooks/useProductFilter.hook';
 
 import type { ProductOptions } from '../Products.types';
 import { Category } from '@/shared/types';
@@ -19,95 +19,32 @@ const ProductFilter: React.FC<ProductFilterProps> = ({
     disable = false,
     className,
 }: ProductFilterProps) => {
-    const resetCategoriesCheckState = useMemo(() => {
-        const initialCategoriesState = new Map<Category, boolean>();
-        Object.values(Category).forEach((category) => {
-            initialCategoriesState.set(category, false);
-        });
-        return initialCategoriesState;
-    }, []);
-    const [categoriesCheckState, setCategoriesCheckState] = useState<Map<Category, boolean>>(resetCategoriesCheckState);
-    const [minimumRating, setMinimumRating] = useState<number>(0);
-    const [minimumPrice, setMinimumPrice] = useState<number>(0);
-    const [maximumPrice, setMaximumPrice] = useState<number>(Infinity);
-    const [title, setTitle] = useState<string | undefined>(undefined);
-    const [brand, setBrand] = useState<string | undefined>(undefined);
-    const [error, setError] = useState<string | undefined>(undefined);
-
-    useEffect(() => {
-        const initializeCategoriesCheckState = () => {
-            const initialCategoriesState = new Map<Category, boolean>();
-            Object.values(Category).forEach((category) => {
-                const isCategoryChecked = initialFilterOptions?.categories?.includes(category) ?? false;
-                initialCategoriesState.set(category, isCategoryChecked);
-            });
-            return initialCategoriesState;
-        };
-
-        setCategoriesCheckState(initializeCategoriesCheckState);
-        setMinimumRating(initialFilterOptions?.minRating ?? 0);
-        setMinimumPrice(initialFilterOptions?.minPrice ?? 0);
-        setMaximumPrice(initialFilterOptions?.maxPrice ?? Infinity);
-        setTitle(initialFilterOptions?.title ?? undefined);
-        setBrand(initialFilterOptions?.brand ?? undefined);
-    }, []);
-
-    const handleConfirmFilter = () => {
-        if (minimumPrice > maximumPrice) {
-            setError('Minimum price cannot be greater than maximum price');
-            return;
-        }
-
-        const updatedProductOptions: Partial<ProductOptions> = {
-            categories: Object.values(Category).filter((category) => categoriesCheckState.get(category)),
-            minRating: minimumRating,
-            minPrice: minimumPrice,
-            maxPrice: maximumPrice !== Infinity ? maximumPrice : undefined,
-            title: title?.trim() !== '' ? title?.trim() : undefined,
-            brand: brand?.trim() !== '' ? brand?.trim() : undefined,
-        };
-
-        if (updatedProductOptions.categories?.length === 0) {
-            updatedProductOptions.categories = undefined;
-        }
-
-        setError(undefined);
-        onFilterChange(updatedProductOptions);
-    };
-
-    const handleResetFilter = () => {
-        setCategoriesCheckState(resetCategoriesCheckState);
-        setMinimumRating(0);
-        setMinimumPrice(0);
-        setMaximumPrice(Infinity);
-        setTitle(undefined);
-        setBrand(undefined);
-        setError(undefined);
-
-        onFilterChange({
-            categories: undefined,
-            minRating: 0,
-            minPrice: 0,
-            maxPrice: undefined,
-            title: undefined,
-            brand: undefined,
-        });
-    };
+    const {
+        categoriesCheckState,
+        minimumRating,
+        minimumPrice,
+        maximumPrice,
+        title,
+        brand,
+        error,
+        handleConfirmFilter,
+        handleResetFilter,
+        setCategoriesCheckState,
+        setMinimumRating,
+        setMinimumPrice,
+        setMaximumPrice,
+        setTitle,
+        setBrand,
+    } = useProductFilter({ initialFilterOptions, onFilterChange });
 
     const handleMinimumRatingChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         const rating = 5 - parseInt(event.target.value);
-
         setMinimumRating(rating);
     };
 
     const handleSelectedCategoriesChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         const category = event.target.value as Category;
-
-        setCategoriesCheckState((prev) => {
-            const updatedCategoriesCheckState = new Map(prev);
-            updatedCategoriesCheckState.set(category, !prev.get(category));
-            return updatedCategoriesCheckState;
-        });
+        setCategoriesCheckState(category);
     };
 
     const handleMinPriceChange = (event: React.ChangeEvent<HTMLInputElement>) => {
